fix: return next list marker as match instead of the current one
This will not make a difference for most lists that always
continue with the same marker. But some lists will return
markers of different lengths (e.g. numbered lists). In those
instances, returning the next marker instead of the previous
one will be useful for correctly adjusting the cursor position.